pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "policygate-capital"
version = "0.1.0"
description = "Deterministic runtime capital governance for autonomous trading systems"
readme = {text = "", content-type = "text/plain"}
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "PolicyGate" },
]
keywords = ["trading", "risk", "governance", "policy", "capital"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "pydantic>=2.0,<3.0",
    "PyYAML>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
